Move processors from `static` source into root so that they can be used by custom sources when needed.
